Drive a real disk from userland
Put a driver under the storage model. A userland walker reads a device id off
the virtio bus, carves the window and the interrupt line for it, starts a driver
for what it found, and checks the sectors while it is still the only process that
can reach them. The kernel learns nothing about disks.

Highlights:
- Add `VirtioBus`, the one process that reads device ids, and `BlockServer`, the
  virtio-blk driver it starts, with barriers at the three points where a
  descriptor changes owner.
- Deliver interrupts as messages on an endpoint (`irqBind`), so one `receive`
  covers both a client and a device, and a driver is a queue rather than a
  process parked in a wait.
- Let a server hold several calls open and answer one by the identity it arrived
  with, so a read can be started and collected instead of blocking a client per
  sector. The server reports how deep one client's own asking got and, separately,
  how many requests were out with the device: the first is a property of the
  client's code, the second is a race a quick disk wins.
- Add `shmPages`, so a server stops believing a client about the size of a
  window it was granted, and `identityAlive`, so it can sweep what a client that
  died left behind.
- Mint the wall clock and the power switch the same way, and hand them to init:
  reading the time needs nothing, setting it is authority, and neither is
  something a process should be able to do by asking politely.
- Widen the caps table to thirty-two slots, which moves `ProcessMetadata` up one
  heap bucket: a window on a device is a capability like any other.
- Retire the demo images. `Child`, `Child2` and `Hello` were written to prove the
  kernel could spawn, share memory and exit, and it does; the Process Server goes
  with them, since its whole interface was an enum naming those two ELFs. The
  boot matrix now ends on the terminal server's line rather than on a demo's
  shared-memory round trip, which is a marker a real service prints.
- Boot the matrix against a real `virtio-blk` drive, and read the queue's
  barriers back out of the linked image, which is the only check a barrier can
  be given.

+++ Sources/ReixKernel/Arch/aarch64/Exceptions/Providers/RXTask/SpawnProcessSyscall.swift
@@ -113,7 +113,7 @@ public struct SpawnProcessSyscall: SyscallProvider {
                     handle: grants[i].sourceHandle,
                     to    : childProcess,
                     slot  : grants[i].targetSlot,
-                    rights: CapRights(rawValue: UInt8(truncatingIfNeeded: grants[i].rights))
+                    rights: CapRights(rawValue: UInt16(truncatingIfNeeded: grants[i].rights))
                 )
             }
